Fix all compilation errors from BigInt migration (forge agent)
The forge agent systematically fixed 29 compilation errors across 16 files
resulting from the chainId and ECDSASignature.v migration from Byte to BigInt.

Fixed patterns:
- Removed redundant BigInt() wrappers around already-BigInt values
- Changed .toByte to BigInt() where constructors expect BigInt
- Updated EvmConfig.chainId from Byte to BigInt
- Fixed all ECDSASignature constructor calls to use (BigInt, BigInt, BigInt)

Files changed:
- BaseETH6XMessages.scala: Removed BigInt() wrappers
- AuthInitiateEcdsaCodec.scala: Changed v.toByte to BigInt(v)
- TestEthBlockServiceWrapper.scala: Removed BigInt() wrapper
- BlockchainConfigForEvm.scala: Changed chainId type to BigInt
- VMServer.scala: Fixed chainId conversion
- ECDSASignatureImplicits.scala: Fixed v value conversions
- EthInfoService.scala: Fixed ECDSASignature constructor
- 9 test files: Updated all ECDSASignature calls to use BigInt

All compilation errors resolved. Ready for testing.

+++ src/main/scala/com/chipprbots/ethereum/jsonrpc/EthInfoService.scala
@@ -175,7 +175,7 @@ class EthInfoService(
       val toAddress = req.tx.to.map(Address.apply)
 
       val tx = LegacyTransaction(0, req.tx.gasPrice, gasLimit, toAddress, req.tx.value, req.tx.data)
-      val fakeSignature = ECDSASignature(0, 0, 0.toByte)
+      val fakeSignature = ECDSASignature(0, 0, 0)
       SignedTransactionWithSender(tx, fakeSignature, fromAddress)
     }
