+++ src/test/scala/com/chipprbots/ethereum/jsonrpc/FukuiiServiceSpec.scala
@@ -65,7 +65,7 @@ class FukuiiServiceSpec
             value = 0,
             payload = ByteString()
           ),
-          signature = ECDSASignature(0, 0, 0.toByte),
+          signature = ECDSASignature(0, 0, 0),
           sender = Address("0x1234")
         )
